Move FrequencyStrategySelector from core package to the CLI package, as it is a CLI-only picocli binding.
